+++ CHANGES.md
@@ -1,7 +1,5 @@
 # Changes
 
-#### cordova-sqlite-storage 5.0.1-dev
-
 #### cordova-sqlite-storage 5.0.0
 
 - avoid incorrect default directory on iOS/macOS - to be extra safe (see <https://github.com/xpbrew/cordova-sqlite-storage/issues/907>)
@@ -14,7 +12,7 @@
 
 #### cordova-sqlite-storage 3.4.1
 
-- SQLite 3.30.1 update from cordova-sqlite-storage-dependencies@2.1.1
+- SQLite 3.31.1 update from cordova-sqlite-storage-dependencies@2.1.1
 
 #### cordova-sqlite-storage 3.4.0